% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marco.R
\name{marco}
\alias{marco}
\title{Hello world}
\usage{
marco()
}
\value{
a hello world message (polo!)
}
\description{
A hello world tester
}
\details{
No, I did not draw this.  It has taken me longer to document this function than it did to copy-paste it from Google (https://www.asciiart.eu/art-and-design/sculptures)
}
\examples{
RRASSLER::marco()
}
\concept{helper}
